calc_ResidEm_grp

Generate pie-charts with the contribution of different within-region population groups to different pollutant emissions by scenario and year

calc_ResidEm_grp(
  db_path = NULL,
  query_path = "./inst/extdata",
  db_name = NULL,
  prj_name,
  scen_name,
  queries = "queries_rhap.xml",
  final_db_year = 2100,
  year,
  pollutant,
  region,
  saveOutput = T,
  pie = T
)

Arguments

db_path
Path to the GCAM database

query_path
Path to the query file

db_name
Name of the GCAM database

prj_name
Name of the rgcam project. This can be an existing project, or, if not, this will be the name

scen_name
Vector names of the GCAM scenarios to be processed

queries
Name of the GCAM query file. The file by default includes the queries required to run rfasst

final_db_year
Final year in the GCAM database (this allows to process databases with user-defined "stop periods")

year
Select the year. 2020:2100

pollutant
Select the pollutant. List of available gases in "rhap::all_pollutants"

region
Select the region. It includes the GCAM 32 geopolitical regions + EU-27 + Global: https://github.com/JGCRI/gcam-doc/blob/gh-pages/overview.md

saveOutput
Writes the emission files. By default=T

pie
Produce the maps. By default=F

Value

Contribution of different groups to pollutant emissions