calc_hap_impacts

Run the model to estimate the health impacts attributable to household air pollution for GCAM scenarios

calc_hap_impacts(
  db_path = NULL,
  query_path = "./inst/extdata",
  db_name = NULL,
  prj_name,
  scen_name,
  queries = "queries_rhap.xml",
  final_db_year = 2100,
  saveOutput = T,
  map = F,
  anim = T,
  HIA_var = "deaths",
  normalized = F,
  by_gr = F
)

Arguments

db_path

Path to the GCAM database

query_path

Path to the query file

db_name

Name of the GCAM database

prj_name

Name of the rgcam project. This can be an existing project, or, if not, this will be the name

scen_name

Vector names of the GCAM scenarios to be processed

queries

Name of the GCAM query file. The file by default includes the queries required to run rfasst

final_db_year

Final year in the GCAM database (this allows to process databases with user-defined "stop periods")

saveOutput

Writes the emission files. By default=T

map

Produce the maps. By default = F

anim

If set to T, produces multi-year animations. By default=T

HIA_var

Health metric to be predicted. c("deaths", "yll", "dalys"). By default = deaths

normalized

Transform the output to "normalized" values. By default = F

by_gr

Estimate damages at group level. Just for illustrative purposes. By default = F

Value

Health impacts attributable to HAP for all the selected years